Fix review issues.
Cherry-picked from: https://gitcode.com/openHiTLS/openhitls/merge_requests/403

+++ codecs/src/decode_chain.c
@@ -108,7 +108,7 @@ void CRYPT_DECODE_PoolFreeCtx(CRYPT_DECODER_PoolCtx *poolCtx)
     BSL_SAL_Free(poolCtx);
 }
 
-static int32_t SetDecodeType(void *val, size_t valLen, const char **targetValue)
+static int32_t SetDecodeType(void *val, int32_t valLen, const char **targetValue)
 {
     if (valLen == 0 || valLen > MAX_CRYPT_DECODE_FORMAT_TYPE_SIZE) {
         BSL_ERR_PUSH_ERROR(CRYPT_INVALID_ARG);
